browser-use-agent/pyproject.toml
[project]
name = "browser-use-agent"
version = "0.1.0"
description = "Browser automation agent using DeepAgents and Ralph Mode"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.1.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "langgraph-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "websockets>=12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
